+++ src/shell.js
@@ -387,6 +387,10 @@ assert(typeof Module['setWindowTitle'] === 'undefined', 'Module.setWindowTitle o
 {{{ makeRemovedModuleAPIAssert('readAsync') }}}
 {{{ makeRemovedModuleAPIAssert('readBinary') }}}
 // TODO: add when SDL2 is fixed {{{ makeRemovedModuleAPIAssert('setWindowTitle') }}}
+{{{ makeRemovedFSAssert('IDBFS') }}}
+{{{ makeRemovedFSAssert('PROXYFS') }}}
+{{{ makeRemovedFSAssert('WORKERFS') }}}
+{{{ makeRemovedFSAssert('NODEFS') }}}
 
 #if USE_PTHREADS
 assert(ENVIRONMENT_IS_WEB || ENVIRONMENT_IS_WORKER, 'Pthreads do not work in non-browser environments yet (need Web Workers, or an alternative to them)');